Run intercept finalizers in their own goroutine
AddInterceptFianlizer is called by stuff outside the state struct; hence
it makes no sense to allow a fianlizer to run holding the lock. If
needed, it can re-acquire it through the state object; the lock is only
protecting the state itself, i.e. the removal of the intercept from the
map.

This also requires some coordination to make sure that the calls into
the state object wait for finalizers to run. There's a new wait()
mechanism to enable this.

Signed-off-by: Jose Cortes <josecortes@datawire.io>